+++ packages/core/src/interpreter.ts
@@ -208,6 +208,10 @@ export class Interpreter<
   public children: Map<string | number, ActorRef<any>> = new Map();
   private forwardTo: Set<string> = new Set();
 
+  private _outgoingQueue: Array<
+    [{ send: (ev: unknown) => void }, unknown]
+  > = [];
+
   // Dev Tools
   private devTools?: any;
 
@@ -324,6 +328,11 @@ export class Interpreter<
       this.options.execute
     ) {
       this.execute(this.state);
+    } else {
+      let item: typeof this._outgoingQueue[number] | undefined;
+      while ((item = this._outgoingQueue.shift())) {
+        item[0].send(item[1]);
+      }
     }
 
     // Update children
@@ -369,6 +378,7 @@ export class Interpreter<
         listener(doneInvoke(this.id, doneData));
       }
       this._stop();
+      this._stopChildren();
     }
   }
   /*
@@ -550,6 +560,15 @@ export class Interpreter<
     });
     return this;
   }
+  private _stopChildren() {
+    // TODO: think about converting those to actions
+    this.children.forEach((child) => {
+      if (isFunction(child.stop)) {
+        child.stop();
+      }
+    });
+    this.children.clear();
+  }
   private _stop() {
     for (const listener of this.listeners) {
       this.listeners.delete(listener);
@@ -623,7 +642,7 @@ export class Interpreter<
           this.state,
           this.state.context,
           _event,
-          exitActions,
+          [exitActions],
           this.machine.config.predictableActionArguments
             ? this._exec
             : undefined,
@@ -658,15 +677,7 @@ export class Interpreter<
       });
 
       this.update(nextState, _event);
-
-      // TODO: think about converting those to actions
-      // Stop all children
-      this.children.forEach((child) => {
-        if (isFunction(child.stop)) {
-          child.stop();
-        }
-      });
-      this.children.clear();
+      this._stopChildren();
 
       registry.free(this.sessionId);
     });
@@ -799,7 +810,8 @@ export class Interpreter<
 
   private sendTo = (
     event: SCXML.Event<AnyEventObject>,
-    to: string | number | ActorRef<any>
+    to: string | number | ActorRef<any>,
+    immediate: boolean
   ) => {
     const isParent =
       this.parent && (to === SpecialTargets.Parent || this.parent.id === to);
@@ -838,16 +850,25 @@ export class Interpreter<
         this.state.done
       ) {
         // Send SCXML events to machines
-        (target as AnyInterpreter).send({
+        const scxmlEvent = {
           ...event,
           name:
             event.name === actionTypes.error ? `${error(this.id)}` : event.name,
           origin: this.sessionId
-        });
+        };
+        if (!immediate && this.machine.config.predictableActionArguments) {
+          this._outgoingQueue.push([target, scxmlEvent]);
+        } else {
+          (target as AnyInterpreter).send(scxmlEvent);
+        }
       }
     } else {
       // Send normal events to other targets
-      target.send(event.data);
+      if (!immediate && this.machine.config.predictableActionArguments) {
+        this._outgoingQueue.push([target, event.data]);
+      } else {
+        target.send(event.data);
+      }
     }
   };
 
@@ -907,7 +928,7 @@ export class Interpreter<
   private defer(sendAction: SendActionObject<TContext, TEvent>): void {
     this.delayedEventsMap[sendAction.id] = this.clock.setTimeout(() => {
       if (sendAction.to) {
-        this.sendTo(sendAction._event, sendAction.to);
+        this.sendTo(sendAction._event, sendAction.to, true);
       } else {
         this.send(
           (sendAction as SendActionObject<TContext, TEvent, TEvent>)._event
@@ -971,7 +992,7 @@ export class Interpreter<
           return;
         } else {
           if (sendAction.to) {
-            this.sendTo(sendAction._event, sendAction.to);
+            this.sendTo(sendAction._event, sendAction.to, _event === initEvent);
           } else {
             this.send(
               (sendAction as SendActionObject<TContext, TEvent, TEvent>)._event